requirements.txt
# Data manipulation and analysis
pandas
numpy
scikit-learn

# Visualization and EDA
matplotlib
seaborn

# Jupyter notebooks for EDA
jupyter

# Modeling
xgboost

# NLP and Transformers
torch
transformers
datasets
peft
shap

# RAG
faiss-cpu

# Interface
streamlit